calculate the pathname each time a file is asked for
the problem to install rameau in /usr/local/share is that the
algorithms are instanciated when rameau is compiled for local use,
saved with cl-store and read (with cl-store) when rameau is
re-compiled to be installed in /usr/local/share. re-defining the
accessor methods to return the pathname for neural files if the slot
is nil.